Back/requirements.txt
# FastAPI 및 웹 서버
fastapi==0.117.1
uvicorn==0.37.0
starlette==0.48.0

# 데이터베이스
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.9

# 인증 및 보안
python-jose==3.5.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4

# HTTP 클라이언트
httpx==0.28.1
requests==2.32.5

# 데이터 검증 및 설정
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic-extra-types==2.10.5
email-validator==2.3.0

# Google AI 및 TTS
google-ai-generativelanguage==0.6.18
google-cloud-texttospeech==2.29.0
google-api-core==2.25.1
google-auth==2.40.3

# LangChain
langchain==0.1.0
langchain-core==0.1.52
langchain-google-genai==2.1.10
langchain-text-splitters==0.3.11
langsmith==0.1.147

# 기타 유틸리티
python-dotenv==1.1.1
pyyaml==6.0.2
pydub==0.25.1
tenacity==9.1.2
rich==14.1.0

# 개발 도구
watchfiles==1.1.0

# Redis
redis==5.0.1

# Airflow API 클라이언트
apache-airflow-client==2.10.0

# WebSocket 지원
websockets==12.0
